Implemented repo showing feature

+++ features/support/repositories.rb
@@ -4,12 +4,13 @@ module RepositoryHelpers
   module_function
 
   def create_repo(name)
-    Grit::Repo.init_bare( File.join(AppConfig.repos_root_dir, "#{name}.git") )
+    repo = Grit::Repo.init( File.join(AppConfig.repos_root_dir, "#{name}.git") )
   end
 
   def destroy_repos
-    FileUtils.rm_rf(AppConfig.repos_root_dir)
+    FileUtils.rm_rf(AppConfig.repos_root_dir) if AppConfig.repos_root_dir
   end
+
 end
 
 World(RepositoryHelpers)
